+++ WatchDogCore/WatchDog/META-INF/MANIFEST.MF
@@ -13,10 +13,7 @@ Bundle-ClassPath: .,
  lib/commons-codec-1.9.jar,
  lib/commons-lang-2.6.jar,
  lib/joda-time-2.8.1.jar,
- lib/mapdb-1.0.5.jar,
- lib/jfreechart-1.0.19-swt.jar,
- lib/jfreechart-1.0.19.jar,
- lib/jcommon-1.0.23.jar
+ lib/mapdb-1.0.5.jar
 Export-Package: com.google.gson,
  com.google.gson.annotations,
  com.google.gson.internal,
@@ -36,21 +33,7 @@ Export-Package: com.google.gson,
  nl.tudelft.watchdog.core.ui.preferences,
  nl.tudelft.watchdog.core.ui.wizards,
  nl.tudelft.watchdog.core.util,
- org.jfree.chart,
- org.jfree.chart.axis,
- org.jfree.chart.plot,
- org.jfree.chart.renderer.category,
- org.jfree.chart.title,
- org.jfree.data.category,
- org.jfree.data.gantt,
- org.jfree.data.general,
- org.jfree.experimental.chart.swt,
- org.jfree.util,
  org.joda.time,
  org.joda.time.base,
  org.joda.time.format,
  org.mapdb
-Import-Package: org.eclipse.swt.events,
- org.eclipse.swt.graphics,
- org.eclipse.swt.layout,
- org.eclipse.swt.widgets
