+++ WatchDogCore/WatchDog/build.properties
@@ -10,7 +10,4 @@ bin.includes = META-INF/,\
                lib/joda-time-2.8.1.jar,\
                lib/gson-2.2.4.jar,\
                lib/mapdb-1.0.5.jar,\
-               pom.xml,\
-               lib/jfreechart-1.0.19-swt.jar,\
-               lib/jfreechart-1.0.19.jar,\
-               lib/jcommon-1.0.23.jar
+               pom.xml
